parsede json integers som int istället för long long, samt lade till GetHash som hashar en godtycklig input stream

+++ MBParsing/MBParsing.cpp
@@ -423,7 +423,7 @@ namespace MBParsing
 		}
 		try
 		{
-			ReturnValue = std::stoi(std::string(ObjectData +IntBegin, ObjectData+ IntEnd));
+			ReturnValue = std::stoll(std::string(ObjectData +IntBegin, ObjectData+ IntEnd));
 			ParseOffset = IntEnd;
 		}
 		catch (const std::exception&)
